Aura/backend/requirements.txt
# Web Framework & API
Flask==2.3.2
flask-restx==1.1.0
gunicorn==20.1.0

# Database & ORM
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6

# Data Processing
pandas==2.0.3

# Async Tasks & Message Broker
celery==5.3.1
redis==4.6.0

# Utilities
python-dotenv==1.0.0
